Check for existence before trying to move to backup dir

+++ install.sh
@@ -25,8 +25,10 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
+    if [ -e ~/.$file ]; then
+        echo "Moving existing dotfile ~/.$file to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
 done
